pyproject.toml
[project]
name = "ringparser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "chardet>=5.2.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.40.0",
    "xmlschema>=4.3.1",
]
